% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ET_forcing-package.R
\docType{package}
\name{ET_forcing-package}
\alias{ET_forcing}
\alias{ET_forcing-package}
\title{ET_forcing: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Dongdong Kong \email{kongdd.sysu@gmail.com} (\href{https://orcid.org/0000-0003-1836-8172}{ORCID})

}
\keyword{internal}
